+++ weatherDB/__init__.py
@@ -9,7 +9,7 @@ import os
 __author__ = "Max Schmit"
 __email__ = "max.schmit@hydrology.uni-freiburg.de"
 __copyright__ = "Copyright 2023, Max Schmit"
-__version__ = "0.0.20"
+__version__ = "0.0.21"
 
 if not ("WEATHERDB_MODULE_INSTALLING" in os.environ \
         and os.environ["WEATHERDB_MODULE_INSTALLING"]=="True"):
@@ -54,6 +54,8 @@ if not ("WEATHERDB_MODULE_INSTALLING" in os.environ \
 
     # import classes
     from . import station, stations
+    from .station import StationN, StationND, StationT, StationET, GroupStation
+    from .stations import StationsN, StationsND, StationsT, StationsET, GroupStations
     try:
         from . import broker
     except PermissionError: